Completed Wassel 4 - Connecting thr Front and Back Ends
Added new funcitionality including displaying comments, adding an upvote button and adding a comment form.  Displaying comments lists out all the comments from the database in the comments section of each post.  The upvote button adds another upvote for that specific article in the database and displays the upvote count on the article.  The comments form allows users to input new comments on the article. There were no bug fixes for this commit.